Refactors and clans up code   - starts work on getting PWM working

+++ altitudeTask.h
@@ -3,13 +3,17 @@
 * altitudeTask.h - prototypes for the altitude task
 */
 
+#include "FreeRTOS.h"
+#include "queue.h"
+
+
 #ifndef ALTITUDETASK_H_
 #define ALTITUDETASK_H_
 
 /*
 * Initializes the altitude task
 */
-extern uint32_t altitudeTaskInit(void);
+extern uint8_t altitudeTaskInit(void);
 
 /**
 * Gets the altitude input queue
